Server security and other improvement: - Redesign FileSystem class to extend base feature and support fully relative, absolute, .., ., in a more safe way - Redesign and rename Path class to File class, feature and security improved - Rename Database to PersistentData, and fix a lot of security issues inside - Make feel secure UnpackURL and UnpackQuery - Redesign DecodeURI and add URI complete support for UnpackURL and UnpackQuery - Fix a lot of security issue on HTTP get file - Allow to use onRead to redirect a HTTP file request to folder view (with a empty returned file) - Improve Util::Environment performance access (becomes static) - Improve Home, CurrentApp and CurrentDir performance access(becomes static) and fix few issues on values computing - Fix a PoolBuffers buffers sorting error (clear size on buffer release)

+++ MonaBase/include/Mona/ServerApplication.h
@@ -48,7 +48,7 @@ private:
 #if defined(_WIN32)
 	static int  __stdcall ConsoleCtrlHandler(unsigned long ctrlType);
 	static void __stdcall ServiceMain(unsigned long argc, char** argv);
-	static void __stdcall ServiceControlHandler(DWORD control);
+	static void __stdcall ServiceControlHandler(unsigned long control);
 
 	bool hasConsole();
 	bool isService();
